pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "django-a2a"
dynamic = ["version"]
authors = [
  { name = "Peter Pak", email = "ppak10@gmail.com" },
]
description = "Django app for handling a2a communication protocol"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]

dependencies = [
  "Django>=4.0",
  "djangorestframework"
]

[project.urls]
Homepage = "https://github.com/ppak10/django-a2a"
Issues = "https://github.com/ppak10/django-a2a/issues"

[tool.hatch.build]
sources = ["src"]

[tool.hatch.build.targets.wheel]
include = [
  "src/**/*",
  "README.md",
]

[tool.hatch.build.targets.sdist]
include = [
  "src/**/*",
  "README.md",
]

[tool.hatch.version]
path = "src/django_a2a/__init__.py"  # or another version file

[tool.black]
line-length = 88
target-version = ["py39"]
